Separate blob creation/update/deletion APIs from other types of images.
Blobs will need some extra APIs that don't apply to regular images.
This change will make it easier to add these extra functionalities
without messing with the rest of the code.
The new BlobImageKey type also adds a tiny bit of type safety to
prevent some mixing blob and non blob data on the same image which
is invalid.